requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
gevent==23.9.1
greenlet==3.0.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0
websocket==0.2.1
websocket-client==1.6.4
Werkzeug==3.0.1
zope.event==5.0
zope.interface==6.1
